Refine recent list and EntryRow for scannability
